# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tienda Perianza API",
        "description": "API REST para administración de productos - Evaluación Técnica",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://143.244.188.9",
            "description": "Servidor VPS"
        }
    ],
    "paths": {
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Consultar productos",
                "operationId": "400eae86648033ed800b2500687783c9",
                "parameters": [
                    {
                        "name": "buscar",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoria",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fecha_inicio",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_fin",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta exitosa"
                    },
                    "401": {
                        "description": "API Key inválida"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "operationId": "942bdb99fe7e9420893ee7c4d43c3856",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "sku",
                                    "categoria",
                                    "precio",
                                    "stock"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "sku": {
                                        "type": "string"
                                    },
                                    "categoria": {
                                        "type": "string"
                                    },
                                    "precio": {
                                        "type": "number"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro guardado"
                    },
                    "401": {
                        "description": "API Key inválida"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/productos/{id}": {
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar producto",
                "operationId": "626bbc89db1d7a3aacc668c8240e9cbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "sku": {
                                        "type": "string"
                                    },
                                    "categoria": {
                                        "type": "string"
                                    },
                                    "precio": {
                                        "type": "number"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro actualizado"
                    },
                    "401": {
                        "description": "API Key inválida"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto (borrado lógico)",
                "operationId": "51834daa2ffb180799292043d0eb86c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro eliminado"
                    },
                    "401": {
                        "description": "API Key inválida"
                    },
                    "422": {
                        "description": "ID requerido"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/productos/export": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Exportar productos a Excel",
                "operationId": "82f28b072ba21fcf14fe8fa0e90740c6",
                "parameters": [
                    {
                        "name": "buscar",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoria",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fecha_inicio",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_fin",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo Excel generado"
                    },
                    "401": {
                        "description": "API Key inválida"
                    },
                    "500": {
                        "description": "Error al exportar"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Productos",
            "description": "Productos"
        }
    ]
}